+++ src/main/java/potemkin/i/yu/Cache.java
@@ -1,3 +1,8 @@
+/*
+ * Все права защищены.
+ * Регулируется лицензией Epam.
+ * 
+ */
 package potemkin.i.yu;
 
 import java.util.Arrays;
@@ -5,14 +10,34 @@ import java.util.Arrays;
 import lombok.extern.slf4j.Slf4j;
 
 /**
- * Параметрорезированный Класс Cache
+ * Параметрорезированный Класс Cache максимальная емкость закрепляется при
+ * создании экземпляра класса через конструктор. Время выполнения операций
+ * линейное. Массив используется для хранения элементов списка, он всекда не
+ * меньше размера списка. При заполненнии емкости списка, при дольнейшей вставке
+ * происходит удаление первого элемента списка и вставка в последнию ячейку.
+ * Емкость списка не может быть увеличина в размере. Данный класс не
+ * синхронизирован.
+ * 
  * 
  * @author Илья Пот
+ * @param <T> тип элементов в этом списке
+ * @since 1
  */
 @Slf4j
 public class Cache<T> {
+	/**
+	 * Массив в котором хронятся элементы класса Cache
+	 */
 	private CacheElement<T>[] cache;
+
+	/**
+	 * Размер массива. Ипользуется для создания размера массива cache.
+	 */
 	private int capacity;
+
+	/**
+	 * nextItem - индекс для следующей вставки элемента
+	 */
 	private int nextItem;
 
 	/**
@@ -99,7 +124,7 @@ public class Cache<T> {
 	}
 
 	/**
-	 * Метод возврата объекта по объекту
+	 * Метод возврата объекта по элементу
 	 * 
 	 * @param element - объект который надо вернуть
 	 * @return возвращаемый объект, если объект не найден вернет null
@@ -158,10 +183,10 @@ public class Cache<T> {
 	public int getCapacity() {
 		return capacity;
 	}
-	
+
 	public int getNextItem() {
 		return nextItem;
-	} 
+	}
 
 	@Override
 	public String toString() {
